Move bridge tests out of bridge module
Extract the bridge test module into a separate test file so `bridge.rs`
stays below the project file-size limit. Replace fixed mtime sleeps in the
IR cache idempotency test with bounded polling that waits for the expected
metadata relation.

Use Oxford spelling for the nextest serialization note in the developers
guide.
